+++ aiter/ops/triton/moe_op_gelu.py
@@ -358,7 +358,7 @@ def _fused_moe_persistent_kernel(
         pid_m, pid_n = pid_grid(tile_id_remapped, num_pid_m, num_pid_n, GROUP_SIZE_M)
 
         # Compute the mask
-        offs_token_id = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
+        offs_token_id = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M).to(tl.int64)
         offs_token = tl.load(sorted_token_ids_ptr + offs_token_id)
         token_mask = offs_token < num_valid_tokens
         off_experts = tl.load(expert_ids_ptr + pid_m).to(tl.int64)
@@ -368,7 +368,7 @@ def _fused_moe_persistent_kernel(
             offs_token[:, None] // top_k * stride_am + offs_k[None, :] * stride_ak
         )
         # Compute the B pointer
-        offs_bn = (pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)) % N
+        offs_bn = (pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N).to(tl.int64)) % N
         b_ptrs = (
             b_ptr
             + off_experts * stride_be
